+++ frameworks/qwik/src/components/Form/Form.tsx
@@ -1,4 +1,9 @@
-import { INTERNAL, type Schema, type SubmitHandler } from '@formisch/core/qwik';
+import {
+  INTERNAL,
+  type Schema,
+  type SubmitEventHandler,
+  type SubmitHandler,
+} from '@formisch/core/qwik';
 import { handleSubmit } from '@formisch/methods/qwik';
 import type { JSXOutput, PropsOf, QRL } from '@qwik.dev/core';
 import { component$, Slot } from '@qwik.dev/core';
@@ -18,7 +23,7 @@ export type FormProps<TSchema extends Schema = Schema> = Omit<
   /**
    * The submit handler called when the form is submitted and validation succeeds.
    */
-  readonly onSubmit$: QRL<SubmitHandler<TSchema>>;
+  readonly onSubmit$: QRL<SubmitHandler<TSchema> | SubmitEventHandler<TSchema>>;
 };
 
 /**
@@ -41,7 +46,7 @@ export const Form = component$(
         ref={(element) => {
           of[INTERNAL].element = element;
         }}
-        onSubmit$={(event) => handleSubmit(of, onSubmit$)(event)}
+        onSubmit$={handleSubmit(of, onSubmit$)}
       >
         <Slot />
       </form>