+++ packages/core/src/types/form.react.ts
@@ -12,9 +12,16 @@ export type {
 } from './form.ts';
 
 /**
- * Submit handler type for React.
+ * Submit handler type.
  */
 export type SubmitHandler<TSchema extends Schema> = (
+  output: v.InferOutput<TSchema>
+) => MaybePromise<unknown>;
+
+/**
+ * Submit event handler type.
+ */
+export type SubmitEventHandler<TSchema extends Schema> = (
   output: v.InferOutput<TSchema>,
   event: FormEvent<HTMLFormElement>
 ) => MaybePromise<unknown>;