+++ website/src/routes/(docs)/core/api/(types)/SubmitEventHandler/properties.ts
@@ -22,4 +22,10 @@ export const properties: Record<string, PropertyProps> = {
       ],
     },
   },
+  event: {
+    type: {
+      type: 'custom',
+      name: 'SubmitEvent',
+    },
+  },
 };